ensures the ui cultures are set in the back office. adds WebSecurity as a dependency for UmbracoContext.

+++ src/umbraco.businesslogic/BasePages/BasePage.cs
@@ -184,7 +184,7 @@ namespace umbraco.BasePages
             var identity = HttpContext.Current.GetCurrentIdentity();
             if (identity == null)
                 return -1;
-            return identity.Id;
+            return (int)identity.Id;
         }
 
         // Added by NH to use with webservices authentications
